+++ src/routes/PrivateRoute/PrivateFoodieRoute.jsx
@@ -4,11 +4,11 @@ import { Route, Redirect } from 'react-router-dom';
 import * as USER_TYPES from '../../constants/userTypes';
 import * as ROUTES from '../../constants/routes';
 
-const Restaurateur = ({ user, component: Component, ...rest }) => (
+const PrivateFoodieRoute = ({ user, component: Component, ...rest }) => (
   <Route
     {...rest}
     render={props =>
-      user && user.isVerified && user.type === USER_TYPES.RESTAURATEUR ? (
+      user && user.isVerified && user.type === USER_TYPES.FOODIE ? (
         <Component {...props} />
       ) : (
         <Redirect to={ROUTES.LANDING} />
@@ -21,4 +21,4 @@ const mapStateToProps = state => {
   return { user: state.auth.user };
 };
 
-export default connect(mapStateToProps)(Restaurateur);
+export default connect(mapStateToProps)(PrivateFoodieRoute);
