+++ src/routes/PrivateRoute/PrivateRestaurateurRoute.jsx
@@ -4,11 +4,11 @@ import { Route, Redirect } from 'react-router-dom';
 import * as USER_TYPES from '../../constants/userTypes';
 import * as ROUTES from '../../constants/routes';
 
-const Foodie = ({ user, component: Component, ...rest }) => (
+const PrivateRestaurateurRoute = ({ user, component: Component, ...rest }) => (
   <Route
     {...rest}
     render={props =>
-      user && user.isVerified && user.type === USER_TYPES.FOODIE ? (
+      user && user.isVerified && user.type === USER_TYPES.RESTAURATEUR ? (
         <Component {...props} />
       ) : (
         <Redirect to={ROUTES.LANDING} />
@@ -21,4 +21,4 @@ const mapStateToProps = state => {
   return { user: state.auth.user };
 };
 
-export default connect(mapStateToProps)(Foodie);
+export default connect(mapStateToProps)(PrivateRestaurateurRoute);
