--- src/routes/private/index.jsx
@@ -1,2 +1,0 @@
-export { default as PrivateRestaurateur } from './Restaurateur';
-export { default as PrivateFoodie } from './Foodie';
