--- src/routes/public/index.jsx
@@ -1,1 +1,0 @@
-export { default as Public } from './Public';
